+++ src/components/back-link-helper/back-link-helper.browser.test.js
@@ -8,7 +8,7 @@ describe('/components/back-link-helper', () => {
   }
 
   describe('When a JS-enabled back link is included on a page', () => {
-    it('should be hidden when referrer is on an invalid domain', async () => {
+    it('should be hidden when referrer is on a different domain', async () => {
       await render(page, withHmrcStylesAndScripts(`
         <a href="#" class="govuk-back-link" data-module="hmrc-back-link">back</a>
       `), {
@@ -17,7 +17,7 @@ describe('/components/back-link-helper', () => {
       expect(await linkDisplayStyle()).toBe('none');
     });
 
-    it('should be shown when referrer is on an allowed domain', async () => {
+    it('should be shown when referrer is on a different domain, but the domain has been allowlisted', async () => {
       await render(page, withHmrcStylesAndScripts(`
         <a href="#" class="govuk-back-link" data-module="hmrc-back-link">back</a>
       `), {