+++ src/components/back-link-helper/back-link-helper.js
@@ -7,8 +7,7 @@ function BackLinkHelper($module, window, document) {
 BackLinkHelper.prototype.init = function init() {
   // do nothing if History API is absent
   if (this.window.history) {
-    // eslint-disable-next-line max-len
-    /* TODO: It remains unclear whether a check for the same domain is necessary for security reasons.
+    /* TODO: Remains unclear whether a check for the same domain is necessary for security reasons.
        There may be user research suggesting considerations regarding the visibility of the
        back link on refresh.
        Currently, a page refresh sets the referer to empty, leading to the back link being hidden
@@ -18,12 +17,11 @@ BackLinkHelper.prototype.init = function init() {
     const shouldHideBackLink = () => {
       const referer = this.document.referrer;
       if (!referer) return true;
-      const referredFromDifferentDomain = referer.indexOf(this.window.location.host) === -1;
-      const referrerNotOnAllowList = !referrerAllowList.some(
-        // eslint-disable-next-line comma-dangle
-        (allowListedDomain) => referer.includes(allowListedDomain)
+      const referredFromDifferentDomain = () => referer.indexOf(this.window.location.host) === -1;
+      const referrerNotOnAllowList = () => !referrerAllowList.some(
+        (allowListedDomain) => referer.includes(allowListedDomain),
       );
-      return referredFromDifferentDomain && referrerNotOnAllowList;
+      return referredFromDifferentDomain() && referrerNotOnAllowList();
     };
 
     // hide the backlink if the referrer is on a different domain or the referrer is not set